Graph 3
This is it. This is the graph I was imagining and more.
I've borrowed some code from another site and done a bunch of
modifications.

It's still a litte messy and I need to remove the old graph code
from graph 1 and 2 but this is beautiful.

A truly interactive and navigatable chart of the mirth channels
in the browser.