Embed graph.html in binary
Using the esc library I was able to compress the ui
directory into the server/ui.go file.

This is done prior to building using the "go generate"
command.  The esc program generates the ui.go file
embeding the assets.

The graph.html file ends up in this ui.go file as a
gzip base64 encoded string.

Because of the size of this file I ran into the same problem
I had when I was serving it off of disk, namely that it
would not all fit into memory as a single string i.e. it
needed to be streamed.

unfortunatly the code generated by esc does not support this
I was able to implement StreamData.go as an extension to
the esc generated code.  (Since the ui.go is generated
dynamically I didn't want to modify it directly.) since this
code is loaded into the server package I was able to write
the FSIoCopy function and the _escStaticFS.IOCopy function
and still have access to the package private members.

I was suprised that once I fixed the small syntax errors this
worked the first time.